+++ st3/mdpopups/st_scheme_template.py
@@ -16,14 +16,14 @@ https://manual.macromates.com/en/language_grammars#naming_conventions
 import sublime
 import re
 from . import version as ver
-from .rgba import RGBA
+from coloraide import util
+from .st_colormod import Color
 from .st_color_scheme_matcher import ColorSchemeMatcher
 import jinja2
 from pygments.formatters import HtmlFormatter
 from collections import OrderedDict
 from .st_clean_css import clean_css
 import copy
-import decimal
 
 NEW_SCHEMES = int(sublime.version()) >= 3150
 
@@ -46,19 +46,156 @@ OLD_DEFAULT_CSS = 'Packages/mdpopups/css/default.css'
 DEFAULT_CSS = 'Packages/mdpopups/mdpopups_css/default.css'
 
 
-def fmt_float(f, p=0):
-    """Set float precision and trim precision zeros."""
+class _Filters:
+    """Color filters."""
 
-    string = str(
-        decimal.Decimal(f).quantize(decimal.Decimal('0.' + ('0' * p) if p > 0 else '0'), decimal.ROUND_HALF_UP)
-    )
+    @staticmethod
+    def colorize(color, deg):
+        """Colorize the color with the given hue."""
 
-    m = re_float_trim.match(string)
-    if m:
-        string = m.group('keep')
-        if m.group('keep2'):
-            string += m.group('keep2')
-    return string
+        h = color.get('hsl.hue')
+        h = util.clamp(deg, 0, 360)
+        color.set('hsl.hue', h)
+
+    @staticmethod
+    def hue(color, deg):
+        """Shift the hue."""
+
+        h = color.get('hsl.hue')
+        h += deg
+        h = color.set('hsl.hue', h % 360)
+
+    @staticmethod
+    def contrast(color, factor):
+        """Adjust contrast."""
+
+        r, g, b = [util.round_half_up(util.clamp(c * 255, 0, 255)) for c in color.coords()]
+        # Algorithm can't handle any thing beyond +/-255 (or a factor from 0 - 2)
+        # Convert factor between (-255, 255)
+        f = (util.clamp(factor, 0.0, 2.0) - 1.0) * 255.0
+        f = (259 * (f + 255)) / (255 * (259 - f))
+
+        # Increase/decrease contrast accordingly.
+        r = util.clamp(util.round_half_up((f * (r - 128)) + 128), 0, 255)
+        g = util.clamp(util.round_half_up((f * (g - 128)) + 128), 0, 255)
+        b = util.clamp(util.round_half_up((f * (b - 128)) + 128), 0, 255)
+        color.red = r
+        color.green = g
+        color.blue = b
+
+    @staticmethod
+    def invert(color):
+        """Invert the color."""
+
+        r, g, b = [util.round_half_up(util.clamp(c * 255, 0, 255)) for c in color.coords()]
+        r ^= 0xFF
+        g ^= 0xFF
+        b ^= 0xFF
+        color.red = r / 255
+        color.green = g / 255
+        color.blue = b / 255
+
+    @staticmethod
+    def saturation(color, factor):
+        """Saturate or unsaturate the color by the given factor."""
+
+        s = color.get('hsl.saturation') / 100.0
+        s = util.clamp(s + factor - 1.0, 0.0, 1.0)
+        color.set('hsl.saturation', s * 100)
+
+    @staticmethod
+    def grayscale(color):
+        """Convert the color with a grayscale filter."""
+
+        luminance = color.luminance()
+        color.red = luminance
+        color.green = luminance
+        color.blue = luminance
+
+    @staticmethod
+    def sepia(color):
+        """Apply a sepia filter to the color."""
+
+        r = util.clamp((color.red * .393) + (color.green * .769) + (color.blue * .189), 0, 1)
+        g = util.clamp((color.red * .349) + (color.green * .686) + (color.blue * .168), 0, 1)
+        b = util.clamp((color.red * .272) + (color.green * .534) + (color.blue * .131), 0, 1)
+        color.red = r
+        color.green = g
+        color.blue = b
+
+    @staticmethod
+    def _get_overage(c):
+        """Get overage."""
+
+        if c < 0.0:
+            o = 0.0 + c
+            c = 0.0
+        elif c > 255.0:
+            o = c - 255.0
+            c = 255.0
+        else:
+            o = 0.0
+        return o, c
+
+    @staticmethod
+    def _distribute_overage(c, o, s):
+        """Distribute overage."""
+
+        channels = len(s)
+        if channels == 0:
+            return c
+        parts = o / len(s)
+        if "r" in s and "g" in s:
+            c = c[0] + parts, c[1] + parts, c[2]
+        elif "r" in s and "b" in s:
+            c = c[0] + parts, c[1], c[2] + parts
+        elif "g" in s and "b" in s:
+            c = c[0], c[1] + parts, c[2] + parts
+        elif "r" in s:
+            c = c[0] + parts, c[1], c[2]
+        elif "g" in s:
+            c = c[0], c[1] + parts, c[2]
+        else:  # "b" in s:
+            c = c[0], c[1], c[2] + parts
+        return c
+
+    @classmethod
+    def brightness(cls, color, factor):
+        """
+        Adjust the brightness by the given factor.
+
+        Brightness is determined by perceived luminance.
+        """
+
+        red, green, blue = [util.round_half_up(util.clamp(c * 255, 0, 255)) for c in color.coords()]
+        channels = ["r", "g", "b"]
+        total_lumes = util.clamp(util.clamp(color.luminance(), 0, 1) * 255 + (255.0 * factor) - 255.0, 0.0, 255.0)
+
+        if total_lumes == 255.0:
+            # white
+            r, g, b = 1, 1, 1
+        elif total_lumes == 0.0:
+            # black
+            r, g, b = 0, 0, 0
+        else:
+            # Adjust Brightness
+            pts = (total_lumes - util.clamp(color.luminance(), 0, 1) * 255)
+            slots = set(channels)
+            components = [float(red) + pts, float(green) + pts, float(blue) + pts]
+            count = 0
+            for c in channels:
+                overage, components[count] = cls._get_overage(components[count])
+                if overage:
+                    slots.remove(c)
+                    components = list(cls._distribute_overage(components, overage, slots))
+                count += 1
+
+            r = util.clamp(util.round_half_up(components[0]), 0, 255) / 255.0
+            g = util.clamp(util.round_half_up(components[1]), 0, 255) / 255.0
+            b = util.clamp(util.round_half_up(components[2]), 0, 255) / 255.0
+        color.red = r
+        color.green = g
+        color.blue = b
 
 
 class SchemeTemplate(object):
@@ -110,8 +247,8 @@ class SchemeTemplate(object):
 
         # Get general theme colors from color scheme file
         self.bground = self.csm.special_colors['background']['color_simulated']
-        rgba = RGBA(self.bground)
-        self.lums = rgba.get_true_luminance()
+        rgba = Color(self.bground)
+        self.lums = rgba.luminance()
         is_dark = self.lums <= LUM_MIDPOINT
         self._variables = {
             "is_dark": is_dark,
@@ -122,7 +259,7 @@ class SchemeTemplate(object):
             "use_pygments": self.use_pygments,
             "default_style": self.default_style
         }
-        self.html_border = rgba.get_rgb()
+        self.html_border = rgba.to_string(hex=True)
         self.fground = self.csm.special_colors['foreground']['color_simulated']
 
     def get_variables(self):
@@ -157,8 +294,8 @@ class SchemeTemplate(object):
 
         if not self.legacy_color_matcher:
             bg = self.get_bg()
-            rgba = RGBA(bg)
-            return rgba.get_true_luminance()
+            rgba = Color(bg)
+            return rgba.luminance()
         else:
             return self.lums
 
@@ -231,9 +368,10 @@ class SchemeTemplate(object):
         try:
             parts = [c.strip('; ') for c in css.split(':')]
             if len(parts) == 2 and parts[0] in ('background-color', 'color'):
-                rgba = RGBA(parts[1] + "%02f" % int(255.0 * max(min(float(factor), 1.0), 0.0)))
-                rgba.apply_alpha(self.get_bg())
-                return '%s: %s; ' % (parts[0], rgba.get_rgb())
+                rgba = Color(parts[1])
+                rgba.alpha = max(min(float(factor), 1.0), 0.0)
+                rgba.overlay(self.get_bg())
+                return '%s: %s; ' % (parts[0], rgba.to_string(hex=True))
         except Exception:
             pass
         return css
@@ -243,9 +381,9 @@ class SchemeTemplate(object):
 
         parts = [c.strip('; ') for c in css.split(':')]
         if len(parts) == 2 and parts[0] in ('background-color', 'color'):
-            rgba = RGBA(parts[1])
-            rgba.colorize(degree)
-            parts[1] = "%s; " % rgba.get_rgb()
+            rgba = Color(parts[1])
+            _Filters.colorize(rgba, degree)
+            parts[1] = "%s; " % rgba.to_string(hex=True)
             return '%s: %s ' % (parts[0], parts[1])
         return css
 
@@ -254,9 +392,9 @@ class SchemeTemplate(object):
 
         parts = [c.strip('; ') for c in css.split(':')]
         if len(parts) == 2 and parts[0] in ('background-color', 'color'):
-            rgba = RGBA(parts[1])
-            rgba.hue(degree)
-            parts[1] = "%s; " % rgba.get_rgb()
+            rgba = Color(parts[1])
+            _Filters.hue(rgba, degree)
+            parts[1] = "%s; " % rgba.to_string(hex=True)
             return '%s: %s ' % (parts[0], parts[1])
         return css
 
@@ -265,9 +403,9 @@ class SchemeTemplate(object):
 
         parts = [c.strip('; ') for c in css.split(':')]
         if len(parts) == 2 and parts[0] in ('background-color', 'color'):
-            rgba = RGBA(parts[1])
-            rgba.invert()
-            parts[1] = "%s; " % rgba.get_rgb()
+            rgba = Color(parts[1])
+            _Filters.invert(rgba)
+            parts[1] = "%s; " % rgba.to_string(hex=True)
             return '%s: %s ' % (parts[0], parts[1])
         return css
 
@@ -276,9 +414,9 @@ class SchemeTemplate(object):
 
         parts = [c.strip('; ') for c in css.split(':')]
         if len(parts) == 2 and parts[0] in ('background-color', 'color'):
-            rgba = RGBA(parts[1])
-            rgba.contrast(factor)
-            parts[1] = "%s; " % rgba.get_rgb()
+            rgba = Color(parts[1])
+            _Filters.contrast(rgba, factor)
+            parts[1] = "%s; " % rgba.to_string(hex=True)
             return '%s: %s ' % (parts[0], parts[1])
         return css
 
@@ -287,9 +425,9 @@ class SchemeTemplate(object):
 
         parts = [c.strip('; ') for c in css.split(':')]
         if len(parts) == 2 and parts[0] in ('background-color', 'color'):
-            rgba = RGBA(parts[1])
-            rgba.saturation(factor)
-            parts[1] = "%s; " % rgba.get_rgb()
+            rgba = Color(parts[1])
+            _Filters.saturation(rgba, factor)
+            parts[1] = "%s; " % rgba.to_string(hex=True)
             return '%s: %s ' % (parts[0], parts[1])
         return css
 
@@ -298,9 +436,9 @@ class SchemeTemplate(object):
 
         parts = [c.strip('; ') for c in css.split(':')]
         if len(parts) == 2 and parts[0] in ('background-color', 'color'):
-            rgba = RGBA(parts[1])
-            rgba.grayscale()
-            parts[1] = "%s; " % rgba.get_rgb()
+            rgba = Color(parts[1])
+            _Filters.grayscale(rgba)
+            parts[1] = "%s; " % rgba.to_string(hex=True)
             return '%s: %s ' % (parts[0], parts[1])
         return css
 
@@ -309,9 +447,9 @@ class SchemeTemplate(object):
 
         parts = [c.strip('; ') for c in css.split(':')]
         if len(parts) == 2 and parts[0] in ('background-color', 'color'):
-            rgba = RGBA(parts[1])
-            rgba.sepia()
-            parts[1] = "%s; " % rgba.get_rgb()
+            rgba = Color(parts[1])
+            _Filters.sepia(rgba)
+            parts[1] = "%s; " % rgba.to_string(hex=True)
             return '%s: %s ' % (parts[0], parts[1])
         return css
 
@@ -320,9 +458,9 @@ class SchemeTemplate(object):
 
         parts = [c.strip('; ') for c in css.split(':')]
         if len(parts) == 2 and parts[0] in ('background-color', 'color'):
-            rgba = RGBA(parts[1])
-            rgba.brightness(factor)
-            parts[1] = "%s; " % rgba.get_rgb()
+            rgba = Color(parts[1])
+            _Filters.brightness(rgba, factor)
+            parts[1] = "%s; " % rgba.to_string(hex=True)
             return '%s: %s ' % (parts[0], parts[1])
         return css