+++ docs/src/markdown/about/changelog.md
@@ -10,6 +10,7 @@
   for any API function (except `syntax_highlight`), it will be ignored.
 - **NEW**: Support latest Markdown and Pymdown Extensions. Must use `markdown.extensions.md_in_html` instead of
   `pymdownx.extrarawhtml`.
+- **NEW**: Requires `coloraide` dependency.
 - **FIX**: Fix exception when using Pygments.
 - **FIX**: Fix extra newlines in code blocks.